Handle lack of shared if a bookmark is public
Looking at the XML data coming back from pinboard.in, it would appear that
if a bookmark is private, then the `shared` value exists in the data and is
set to "no"; if it's set in the interface on the website to public (private
isn't ticked), then the `shared` value isn't in the data at all.

So checking if shared is "yes" means every bookmark will always appear to be
private as seen via this library.